requirements.txt
accelerate==0.34.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
axial_positional_embedding==0.2.1
certifi==2024.8.30
charset-normalizer==3.3.2
chronos==0.3
click==8.1.7
CoLT5-attention==0.11.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
docutils==0.21.2
einops==0.8.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gluonts==0.15.1
huggingface-hub==0.25.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.7
local-attention==1.9.15
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
packaging==24.1
pandas==2.2.3
pillow==10.2.0
positional-encodings==6.0.3
product_key_memory==0.2.11
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
reformer-pytorch==1.4.4
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
statistics==1.0.3.5
sympy==1.12
tensorboardx==2.6.2.2
threadpoolctl==3.5.0
tokenizers==0.20.0
toolz==0.12.1
torch==2.4.1
torchaudio==2.4.1
torchvision
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typer==0.12.5
typer-config==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
